fix: env var assignement problem using ??
Null coalescence operator ?? works on 'null' value, undefined variable, and getenv() returns false or something so that variable is defined with an empty but not undefined value.
So using a ternary operator works better on using getenv() function. We could use $_ENV but env var could be found in $_ENV or $_SERVER depending which is the
source of the environment variable, it seems more reliable to use getenv() then.

This is a bit more verbose way to write it but works every times.

Signed-off-by: Romain Forlot <rforlot@yahoo.com>